Refactor APU implementation and enhance audio processing in Xbox 360 emulator
- Consolidated APU source files to always include the stub and xma_decoder for improved audio handling.
- Removed legacy APU class implementation from xma_decoder files to streamline codebase.
- Updated logging macros in xma_decoder for better debugging output.
- Enhanced CPU instruction handling for extended operations and atomic instructions.
- Removed outdated APU integration tests to focus on core functionality.

+++ native/src/cpu/xenon/cpu.h
@@ -155,7 +155,7 @@ struct ThreadContext {
 struct DecodedInst {
     u32 raw;            // Raw instruction word
     u8 opcode;          // Primary opcode (bits 0-5)
-    u8 xo;              // Extended opcode
+    u16 xo;             // Extended opcode (needs >8 bits for values like divd=489)
     u8 rd, rs, ra, rb;  // Register operands
     u8 rc;              // Record bit
     s16 simm;           // Signed immediate
